fix: refactor hook for  modal component
