+++ src/pages/Pools/components/WithdrawLiquidityModal/WithdrawLiquidityModal.tsx
@@ -1,30 +1,20 @@
-import { CW20_DECIMALS, ORAI } from '@oraichain/oraidex-common/build/constant';
 import { toAmount } from '@oraichain/oraidex-common/build/helper';
 import { ReactComponent as CloseIcon } from 'assets/icons/ic_close_modal.svg';
 import cn from 'classnames/bind';
 import { Button } from 'components/Button';
 import Loader from 'components/Loader';
 import Modal from 'components/Modal';
-import { TToastType, displayToast } from 'components/Toasts/Toast';
 import TokenBalance from 'components/TokenBalance';
-import { network } from 'config/networks';
-import { handleCheckAddress, handleErrorTransaction } from 'helper';
 import useConfigReducer from 'hooks/useConfigReducer';
-import CosmJs from 'libs/cosmjs';
-import { useGetPairInfo } from 'pages/Pools/hooks/useGetPairInfo';
-import { useGetPoolDetail } from 'pages/Pools/hooks/useGetPoolDetail';
-import { FC, useState } from 'react';
-import { useSelector } from 'react-redux';
-import { useParams } from 'react-router-dom';
-import { Type, generateContractMessages } from 'rest/api';
-import { RootState } from 'store/configure';
+import { useWithdrawLP } from 'pages/Pools/hooks/useWithdrawLP';
+import { FC } from 'react';
+import { PoolModalProps } from 'types/pool';
 import InputWithOptionPercent from '../InputWithOptionPercent';
-import { ModalProps } from '../MyPoolInfo/type';
 import styles from './WithdrawLiquidityModal.module.scss';
 
 const cx = cn.bind(styles);
 
-export const WithdrawLiquidityModal: FC<ModalProps> = ({
+export const WithdrawLiquidityModal: FC<PoolModalProps> = ({
   isOpen,
   close,
   open,
@@ -33,72 +23,28 @@ export const WithdrawLiquidityModal: FC<ModalProps> = ({
   myLpBalance
 }) => {
   const [theme] = useConfigReducer('theme');
-  let { poolUrl } = useParams();
-  const poolDetail = useGetPoolDetail({ pairDenoms: poolUrl });
-
-  const { token1, token2, info: pairInfoData } = poolDetail;
-  const { lpTokenInfoData, pairAmountInfoData } = useGetPairInfo(poolDetail);
-  const lpPools = useSelector((state: RootState) => state.token.lpPools);
-  const [chosenWithdrawPercent, setChosenWithdrawPercent] = useState(-1);
-  const [lpAmountBurn, setLpAmountBurn] = useState<bigint | null>(null);
-  const [actionLoading, setActionLoading] = useState(false);
-  const [isFocus, setIsFocus] = useState(true);
-
-  const lpTokenBalance = BigInt(pairInfoData ? lpPools[pairInfoData?.liquidityAddr]?.balance ?? '0' : 0);
-  const token1Amount = BigInt(pairAmountInfoData?.token1Amount || 0);
-  const token2Amount = BigInt(pairAmountInfoData?.token2Amount || 0);
-
-  const onChangeWithdrawPercent = (option: number) => {
-    setLpAmountBurn((toAmount(option, CW20_DECIMALS) * lpTokenBalance) / BigInt(100000000));
-  };
-
-  const handleWithdrawLiquidity = async (amount: string) => {
-    if (!pairInfoData) return;
-    setActionLoading(true);
-    displayToast(TToastType.TX_BROADCASTING);
-    try {
-      const oraiAddress = await handleCheckAddress('Oraichain');
-
-      const msg = generateContractMessages({
-        type: Type.WITHDRAW,
-        sender: oraiAddress,
-        lpAddr: lpTokenInfoData!.contractAddress!,
-        amount,
-        pair: pairInfoData.pairAddr
-      });
-
-      const result = await CosmJs.execute({
-        address: msg.contractAddress,
-        walletAddr: oraiAddress,
-        handleMsg: msg.msg,
-        gasAmount: { denom: ORAI, amount: '0' },
-        funds: msg.funds
-      });
-
-      if (result) {
-        displayToast(TToastType.TX_SUCCESSFUL, {
-          customLink: `${network.explorer}/txs/${result.transactionHash}`
-        });
-        setLpAmountBurn(0n);
-        onLiquidityChange(-lpAmountBurnUsdt);
-      }
-    } catch (error) {
-      console.log('error in Withdraw Liquidity: ', error);
-      handleErrorTransaction(error);
-    } finally {
-      setActionLoading(false);
-    }
-  };
+  const {
+    actionLoading,
+    token1,
+    token2,
+    pairInfoData,
+    lpTokenBalance,
+    lpTokenInfoData,
+    lp1BurnAmount,
+    lp2BurnAmount,
+    lpAmountBurnUsdt,
+    lpAmountBurn,
+    setLpAmountBurn,
+    handleWithdrawLiquidity
+  } = useWithdrawLP({
+    myLpBalance,
+    myLpUsdt,
+    onLiquidityChange
+  });
 
   const Token1Icon = theme === 'light' ? token1?.IconLight || token1?.Icon : token1?.Icon;
   const Token2Icon = theme === 'light' ? token2?.IconLight || token2?.Icon : token2?.Icon;
 
-  const totalSupply = BigInt(lpTokenInfoData?.total_supply || 0);
-  const lp1BurnAmount =
-    totalSupply === BigInt(0) || !lpAmountBurn ? BigInt(0) : (token1Amount * BigInt(lpAmountBurn)) / totalSupply;
-  const lp2BurnAmount =
-    totalSupply === BigInt(0) || !lpAmountBurn ? BigInt(0) : (token2Amount * BigInt(lpAmountBurn)) / totalSupply;
-  const lpAmountBurnUsdt = !myLpBalance ? 0 : (Number(lpAmountBurn) / Number(myLpBalance)) * Number(myLpUsdt);
   return (
     <Modal isOpen={isOpen} close={close} open={open} isCloseBtn={false} className={cx('modal')}>
       <div className={cx('container', theme)}>
